Fuchsify better
The idea of constructing balances with only one basis vector of
the eigenspace did not turn out to be that great: such balances do
not lower the residue rank automatically, but only by accident,
and these accidents do not occur that often for bigger matrices,
as it turned out. This part is no more.

Balancing between points of the same Poincare rank on the other
hand, while doesn't guarantee progress automatically (it can
spoil more than it reduces), in a few important cases turned out
to produce much better reductions overall. This part is added.